Move favicons to site root
This will apply to other pages as well, not just Zola articles
containing <link rel="icon">. Additionally this fixes the internal paths
in the webmanifest, and adds a link to it.

I have no idea what I'm doing. And even before this change, adding a
page to my Android home screen would generate a hi-res icon.